requirements.txt
innertube~=2.1.16
opencv-python~=4.10.0.84
pillow~=10.4.0